src/main/java/io/vitalir/vitalirspring/features/user/domain/RegistrationServiceImpl.java
package io.vitalir.vitalirspring.features.user.domain;

import io.vitalir.vitalirspring.features.user.domain.model.Role;
import io.vitalir.vitalirspring.features.user.domain.model.User;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

@Service
public class RegistrationServiceImpl implements RegistrationService {

    private final UserRepository userRepository;
    private final PasswordEncoder passwordEncoder;

    public RegistrationServiceImpl(UserRepository userRepository, PasswordEncoder passwordEncoder) {
        this.userRepository = userRepository;
        this.passwordEncoder = passwordEncoder;
    }

    @Override
    public boolean register(String email, String password) {
        var existingUser = userRepository.getUserByEmail(email);
        if (existingUser.isPresent()) {
            return false;
        }
        var newUser = new User(email, passwordEncoder.encode(password), Role.USER);
        userRepository.save(newUser);
        return true;
    }
}
